Final design & implementation: Physical memory is mapped at 4K granularity and in chunks of {1G,2M,4K}, wherever applicable. In order to permit the clearing loop to skip reserved page frames (e.g. page tables or image pages:) a) CRA_PE_PRESENT is cleared in their PTEs, making them inaccessible, and b) a special-cased #PF exception handler adjusts %{rcx,rdi} accordingly    and restarts or skips the REP STOSQ instruction if %rcx is still >0.
As this permits leveraging the information already present in the
form of the page tables, no data structures need to be shared with
the self-contained environment (subr_clear.c) and, as such, mapped.

Processing physical memory in chunks of {1G,2M,4K} additionally allows
for compacting the page tables towards {1G,2M} granularity depending on
CPU support, and will be implemented in future commits along w/ SMP support
in the clearing loop.